+++ ci/travis/script.sh
@@ -40,7 +40,7 @@ ${TRAVIS_BUILD_DIR}/build/out/osqp_tester
 
 if [[ $TRAVIS_OS_NAME == "linux" ]]; then
     # Check memory with valgrind
-    valgrind --suppressions=${TRAVIS_BUILD_DIR}/.valgrind-suppress.supp --leak-check=full --gen-suppressions=all --error-exitcode=42 ${TRAVIS_BUILD_DIR}/build/out/osqp_tester
+    valgrind --suppressions=${TRAVIS_BUILD_DIR}/.valgrind-suppress.supp --leak-check=full --gen-suppressions=all --track-origins=yes --error-exitcode=42 ${TRAVIS_BUILD_DIR}/build/out/osqp_tester
     # Perform code coverage
     cd ${TRAVIS_BUILD_DIR}/build
     lcov --directory . --capture -o coverage.info # capture coverage info
@@ -82,3 +82,32 @@ cd build
 cmake -G "Unix Makefiles" -DPRINTING=OFF -DUNITTESTS=ON ..
 make
 ${TRAVIS_BUILD_DIR}/build/out/osqp_tester
+
+
+# Test custom memory management
+# ---------------------------------------------------
+
+echo "Test OSQP custom allocators"
+cd ${TRAVIS_BUILD_DIR}
+rm -rf build
+mkdir build
+cd build
+cmake -DUNITTESTS=OFF \
+    -DOSQP_CUSTOM_MEMORY=${TRAVIS_BUILD_DIR}/tests/custom_memory/custom_memory.h \
+    ..
+#make the static library first: it will be missing symbols
+#for the custom memory functions
+make osqpstatic
+#compile the custom memory management functions
+g++ -c ${TRAVIS_BUILD_DIR}/tests/custom_memory/custom_memory.c \
+    -o ${TRAVIS_BUILD_DIR}/build/custom_memory.o
+#for testing purposes, just add this to osqp
+#static library.   For real applications, users
+#may prefer to keep it separate and link both
+#to their application
+ar -rcs ${TRAVIS_BUILD_DIR}/build/out/libosqp.a \
+        ${TRAVIS_BUILD_DIR}/build/custom_memory.o
+
+#now make the rest of the demos and test
+make osqp_demo
+${TRAVIS_BUILD_DIR}/build/out/osqp_demo