+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Unreleased
 ----------------
 * Added meaningful return values to internal functions. Changed syntax of `osqp_setup` function. It now returns an exitflag.
-* Custom memory allocators via cmake and the configure file
-
+* `osqp_setup` function requires `P` to be upper triangular. It returns a nonzero exitflag otherwise.
+* Custom memory allocators via cmake and the configure file.
 
 
 Version 0.5.0 (10 December 2018)